libfs_avb: Disable dm-verity when AVB is permissive
When the bootloader is unlocked (i.e. AVB is permissive), enforcing
dm-verity on system partitions is meaningless because the bootloader
doesn't enforce verification on the root of the high-level verified boot
chain: the kernel. As a result, mounting system partitions with
dm-verity (hashtree verification) is futile when the code performing
verification has not been verified in the first place; users can also
disable dm-verity manually by flashing vbmeta with `fastboot flash
--disable-verity vbmeta vbmeta.img`.

For user and developer convenience, disable dm-verity automatically when
the bootloader is unlocked by checking for permissive AVB. This makes it
possible to ship enforcing vbmeta images for security-conscious users to
lock their bootloader and reap the benefits of verified boot, while
still allowing users with unlocked bootloaders to modify system
partitions.

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Change-Id: I8c27047a3f75a466eb10a761bc7600b11dbe7653

Former-commit-id: 0076631075198c5bc92f7e7cdd5a7d962a0e5c6e
Former-commit-id: a0b05f3f79da1742677fea22f793ec9f69dc7a0d
Former-commit-id: 7223ac3fa93e86b94f4fa681748ca77e24ce722d
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>