Make public.libraries.android.txt available to filesystem
public.libraries.android.txt was not installable as it was used only
from microdroid. However, this module can also be used from system
image, as there is no difference from the content. This change marks the
module as no_full_install, so it can be installed in the system image
defined in Android.bp

Bug: 343357947
Test: AOSP CF build succeeded
Change-Id: I15fd47b173b5b718972a6ee18b3aecb58826909c

Former-commit-id: ed92eabfbb15bbe6d9e6efe55edc57ef7eb5f6f1